fCoh3d8 subelem major update
this subelem no longer has an intact status. Once it is allocated in the
fCoh3d8 element, it should already has the top_edge_status array
available and it should already be able to be partitioned into sub
elems. So, the changes are:
- remove pstat component
- remove update top edge status procedure
- include top_edge_status as an arg. of set procedure, and check its
validity there
- in integration, go straight to partition element if not yet done, then
update internal nodes and integrate assemble subelem. no need to
separate cases with different pstat.

In short, fCoh3d8 subelem is NOT used for intact state, it is used when
it already can be partitioned.