fCoh3d8 element module compiled
blocked by a newly learnt problem:

array section arg. with vector indexing is only for intent in actual
arg. CANNOT be used for intent inout or out

so, MUST be careful when passing nodes and edges arg. to f- elements, as
they will most likely be intent inout. Create local copies first and
pass the copies in.