Major updates to include theta1&2 in delamination...
Rename:
coh3d6 -> delam6
coh3d8 -> delam8
fCoh3d8sub -> fDelam8sub
fCoh3d8 -> fDelam8

All delam elements are assumed to be on/parallel to the x-y plane
(checks of this assumption are included in delam6 and delam8 elements)

theta1 and theta2 are included in the integration procedure of all delam
elements
- delam6/8 element select the theta which is most favorable for
delamination (the theta along which the ujump is larger)

- fBrickLam element is modified to pass top&bot ply angles to fDelam8
element

Summary:
- Now, the assumed working coordinate system of all elements is x-y-z,
where z is the normal. All fBrickLam elements (and hence fBrick,
fDelam8, fDelam8sub) are parallel to x-y plane
- Separate base cohesive element subroutines for delamination and matrix
crack: delam6/8 and cohCrack, respectively
- delam6/8 has theta1 and theta2 as inputs, cohCrack does not. theta1
and theta2 are top & bot ply fibre angles of the interface.
- cohCrack fibre angles is edge1-2
- fBrick element must partition in such a way that the cohCrack
subelem's edge1-2 is along the fibre direction of the ply.