A serious bug of cohesive elements fixed
Previously, the determinant of jacobian of cohesive elements were
NEGATIVE!!

It is due to the wrong Qmatrix that I used (modified from earilier one
when trying to switch the transverse and longitudinal directions)

Also, in delam coh elems, the normal direction may not be [0,0,1], it
may well be [0,0,-1] !! the right way is still to compute normal =
edge1-2 * edge1-4 , then select tangent1 as the longitudinal direction
along the fibre, and compute tangent2 as normal * tangent1

In coh8 elems, midcoords must be rotated by Qmatrix and then compute the
det of jacob at each integration point. Be careful that the rotation
must be applied to the ORIGINAL midcoords array. So a midcoordsQ array
is added in coh8delam elem to represent the rotated midcoords array at
each intg point, and this array is used for det of jacob calculation